# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ExperimentCompareResponse": {
        "additionalProperties": false,
        "description": "Experiment comparison table.",
        "properties": {
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "metric",
          "rows"
        ],
        "title": "ExperimentCompareResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HumanResponseIn": {
        "additionalProperties": false,
        "description": "Single forced-choice trial.",
        "properties": {
          "choice": {
            "title": "Choice",
            "type": "string"
          },
          "clip_id": {
            "title": "Clip Id",
            "type": "string"
          },
          "compression_status": {
            "default": "clean",
            "title": "Compression Status",
            "type": "string"
          },
          "confidence_1_5": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Confidence 1 5",
            "type": "integer"
          },
          "language": {
            "default": "hi",
            "title": "Language",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "string"
          },
          "response_ms": {
            "minimum": 0.0,
            "title": "Response Ms",
            "type": "integer"
          }
        },
        "required": [
          "participant_id",
          "clip_id",
          "choice",
          "confidence_1_5",
          "response_ms"
        ],
        "title": "HumanResponseIn",
        "type": "object"
      },
...
      "ParticipantCreate": {
        "additionalProperties": false,
        "description": "Volunteer registration (anonymous).",
        "properties": {
          "fluency_self_report": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Fluency Self Report",
            "type": "string"
          }
        },
        "required": [
          "fluency_self_report"
        ],
        "title": "ParticipantCreate",
        "type": "object"
      },
      "ParticipantResponse": {
        "additionalProperties": false,
        "description": "Anonymous participant payload.",
        "properties": {
          "clip_ids": {
            "items": {
              "type": "string"
            },
            "title": "Clip Ids",
            "type": "array"
          },
          "fluency_self_report": {
            "title": "Fluency Self Report",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "string"
          }
        },
        "required": [
          "participant_id",
          "fluency_self_report",
          "clip_ids"
        ],
        "title": "ParticipantResponse",
        "type": "object"
      },
...
      "ReportResponse": {
        "additionalProperties": false,
        "description": "Downloadable report pointer.",
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "report_markdown": {
            "title": "Report Markdown",
            "type": "string"
          }
        },
        "required": [
          "report_markdown",
          "experiment_id"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cross-lingual, compression-robust AI-voice detection for Hindi, Marathi, and Tamil with calibrated reliability.",
    "title": "VaaniQ",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/clips/{clip_id}": {
      "get": {
        "description": "Clip metadata for study UI.",
        "operationId": "dataset_clip_meta_api_v1_datasets_clips__clip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dataset Clip Meta Api V1 Datasets Clips  Clip Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Clip Meta",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/clips/{clip_id}/audio": {
      "get": {
        "description": "Stream a demo-corpus WAV for explorer / human-study playback.",
        "operationId": "dataset_clip_audio_api_v1_datasets_clips__clip_id__audio_get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Clip Audio",
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/v1/experiments/compare": {
      "get": {
        "description": "Compare stored experiments on a named metric.",
        "operationId": "compare_experiments_api_v1_experiments_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "eer",
              "title": "Metric",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentCompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Experiments",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/report": {
      "get": {
        "description": "Downloadable evaluation report markdown.",
        "operationId": "experiment_report_api_v1_experiments_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "experiment_id",
            "required": false,
            "schema": {
              "default": "demo",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment Report",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/search": {
      "get": {
        "description": "Search the research catalogue.",
        "operationId": "search_experiments_api_v1_experiments_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "model_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          },
          {
            "in": "query",
            "name": "rq_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rq Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Search Experiments Api V1 Experiments Search Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Experiments",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/human-study/register": {
      "post": {
        "description": "Register an anonymous volunteer and assign balanced clips (RQ5).",
        "operationId": "register_api_v1_human_study_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "human_study"
        ]
      }
    },
...
    "/api/v1/human-study/response": {
      "post": {
        "description": "Record a timed real/fake judgement with 1-5 confidence.",
        "operationId": "submit_response_api_v1_human_study_response_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanResponseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Submit Response Api V1 Human Study Response Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Response",
        "tags": [
          "human_study"
        ]
      }
    },
...
  }
}